Add two other recipes and modify index.html file
Two new recipes were added under the /recipe directory with their two
images in the /img directory.
Changes were made in the index file in order to link the two new pages
with the homepage.
README file modified in order to add the citations of the authors of the
images and add links to licences